Move price getters file to the lib folder
